Added Models for Web App Client Interface Added ToolControllerApi

+++ YourProjectWebApp/Models/Patron.cs
@@ -7,7 +7,7 @@ namespace YourProjectWebService.Models
 {
     public class Patron
     {
-        public long PatronId { get; set; }
+        public int PatronId { get; set; }
         public string PatronName { get; set; }
         public bool IsGroup { get; set; }
     }
